Macro pretty_assertions::assert_eq

macro_rules! assert_eq {
    ($left:expr, $right:expr$(,)?) => { ... };
    ($left:expr, $right:expr, $($arg:tt)*) => { ... };
    (@ $left:expr, $right:expr, $maybe_colon:expr, $($arg:tt)*) => { ... };
}

Asserts that two expressions are equal to each other (using PartialEq).

On panic, this macro will print a diff derived from Debug representation of each value.

This is a drop in replacement for core::assert_eq!. You can provide a custom panic message if desired.

Examples

use pretty_assertions::assert_eq;

let a = 3;
let b = 1 + 2;
assert_eq!(a, b);

assert_eq!(a, b, "we are testing addition with {} and {}", a, b);