Macro pretty_assertions::assert_str_eq

macro_rules! assert_str_eq {
    ($left:expr, $right:expr$(,)?) => { ... };
    ($left:expr, $right:expr, $($arg:tt)*) => { ... };
    (@ $left:expr, $right:expr, $maybe_semicolon:expr, $($arg:tt)*) => { ... };
}

Asserts that two expressions are equal to each other (using PartialEq).

On panic, this macro will print a diff derived from each value’s str representation. See StrComparison for further details.

This is a drop in replacement for core::assert_eq!. You can provide a custom panic message if desired.

Examples

use pretty_assertions::assert_str_eq;

let a = "foo\nbar";
let b = ["foo", "bar"].join("\n");
assert_str_eq!(a, b);

assert_str_eq!(a, b, "we are testing concatenation with {} and {}", a, b);